pyproject.toml
[tool.poetry]
name = "metaprograming"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
bs4 = "^0.0.1"
python-docx = "0.8.10"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
